feat: track medicine quantities to one decimal place
Quantities (Inventory, InventoryAdjustment, Transaction) move from
Integer to BigDecimal scale-1 (NUMERIC(10,1)), letting stock be
tracked in fractional units instead of whole numbers only. Submitted
quantities with more than one decimal place are rounded HALF_UP
rather than rejected. Money amounts stay whole rupees everywhere —
every price x quantity computation rounds to the nearest integer.

Covers entities, DTOs, services, controllers, DB migration
(DataMigrationService), frontend forms/display (step=0.1, toFixed(1)),
and test suites (backend, frontend, e2e) end to end.

Also fixes two latent DataMigrationService bugs surfaced by this
change: widenInventoryTypeColumns() and the new quantity-widening
migration were re-running ALTER COLUMN TYPE unconditionally on every
startup, even when the column already had the target type. Postgres
still bumps the relation's cache-invalidation counter on a no-op
ALTER, which broke an already-prepared statement on a pooled
Hibernate connection right after seeding ("cached plan must not
change result type" on the first /api/inventory/adjust call). Both
migrations now check the live column type first and skip when
already correct.

+++ backend/src/main/java/com/pharma/inventory/controller/TransactionController.java
@@ -18,6 +18,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -45,7 +46,7 @@ public class TransactionController {
     @PreAuthorize("hasRole('USER')")
     public ResponseEntity<TransactionResponse> submit(
             @RequestParam("medicineId") Long medicineId,
-            @RequestParam("quantity") Integer quantity,
+            @RequestParam("quantity") BigDecimal quantity,
             @RequestParam("notes") String notes,
             @RequestParam(value = "screenshots", required = false) List<MultipartFile> screenshots,
             @RequestParam(value = "pricePerUnit", required = false) Integer pricePerUnit,
@@ -130,7 +131,7 @@ public class TransactionController {
 
     // ── Assembly helper (no logic — only construction) ─────────────────
 
-    private TransactionRequest buildRequest(Long medicineId, Integer quantity,
+    private TransactionRequest buildRequest(Long medicineId, BigDecimal quantity,
                                              String notes, List<MultipartFile> screenshots,
                                              Integer pricePerUnit, String inventoryType)
             throws IOException {
